added button ADD COMMENT on the task view and implemented logic for that

+++ src/main/java/com/example/demo/service/ProjectServiceImpl.java
@@ -114,7 +114,7 @@ public class ProjectServiceImpl{
     }
 
     public void assignUserToProject(Long userId, Long projectId){
-        User user = userRepository.findById(userId.intValue())
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(userId));
 
         Project project = projectRepository.findById(projectId)
@@ -125,7 +125,7 @@ public class ProjectServiceImpl{
     }
 
     public void removeUserFromProject(Long userId, Long projectId){
-        User user = userRepository.findById(userId.intValue())
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(userId));
 
         Project project = projectRepository.findById(projectId)
